+++ modules/chain-grabber/src/main/scala/org/ergoplatform/explorer/indexer/Application.scala
@@ -7,7 +7,7 @@ import fs2.Stream
 import io.chrisdavenport.log4cats.SelfAwareStructuredLogger
 import io.chrisdavenport.log4cats.slf4j.Slf4jLogger
 import monix.eval.{Task, TaskApp}
-import org.ergoplatform.explorer.clients.ergo.ErgoNetworkClient
+import org.ergoplatform.explorer.services.ErgoNetwork
 import org.ergoplatform.explorer.db.{DoobieTrans, Trans}
 import org.ergoplatform.explorer.indexer.processes.{ChainIndexer, EpochsIndexer}
 import org.ergoplatform.explorer.settings.pureConfigInstances._
@@ -30,14 +30,16 @@ object Application extends TaskApp {
   def run(args: List[String]): Task[ExitCode] =
     resources(args.headOption).use[Task, ExitCode] { case (logger, settings, client, trans) =>
       logger.info("Starting Indexers ..") >>
-        ErgoNetworkClient[Task](client, settings.network)
+        ErgoNetwork[Task](client, settings.network)
           .flatMap { ns =>
             mkProgram(ns, settings, trans).compile.drain as ExitCode.Success
           }
           .guarantee(logger.info("Stopping Chain Indexer ..."))
     }
 
-  private def resources(configPathOpt: Option[String]): Resource[Task, (SelfAwareStructuredLogger[Task], IndexerSettings, Client[Task], Trans[ConnectionIO, Task])] =
+  private def resources(
+    configPathOpt: Option[String]
+  ): Resource[Task, (SelfAwareStructuredLogger[Task], IndexerSettings, Client[Task], Trans[ConnectionIO, Task])] =
     for {
       logger   <- Resource.eval(Slf4jLogger.create)
       settings <- Resource.eval(IndexerSettings.load[Task](configPathOpt))
@@ -47,15 +49,15 @@ object Application extends TaskApp {
     } yield (logger, settings, client, trans)
 
   private def mkProgram(
-    ns: ErgoNetworkClient[Task],
+    network: ErgoNetwork[Task],
     settings: IndexerSettings,
     trans: ConnectionIO Trans Task
   ): Stream[Task, Unit] =
     Stream
       .emits(
         List(
-          Stream.eval(ChainIndexer[Task, ConnectionIO](settings, ns)(trans)).flatMap(_.run),
-          Stream.eval(EpochsIndexer[Task, ConnectionIO](settings, ns)(trans)).flatMap(_.run)
+          Stream.eval(ChainIndexer[Task, ConnectionIO](settings, network)(trans)).flatMap(_.run),
+          Stream.eval(EpochsIndexer[Task, ConnectionIO](settings, network)(trans)).flatMap(_.run)
         )
       )
       .parJoinUnbounded
