+++ modules/chain-grabber/src/main/scala/org/ergoplatform/explorer/indexer/processes/ChainIndexer.scala
@@ -12,7 +12,7 @@ import fs2.Stream
 import mouse.anyf._
 import org.ergoplatform.explorer.Err.ProcessingErr.{InconsistentNodeView, NoBlocksWritten}
 import org.ergoplatform.explorer.Id
-import org.ergoplatform.explorer.clients.ergo.ErgoNetworkClient
+import org.ergoplatform.explorer.services.ErgoNetwork
 import org.ergoplatform.explorer.db.Trans
 import org.ergoplatform.explorer.db.algebra.LiftConnectionIO
 import org.ergoplatform.explorer.db.models.{BlockStats, Header}
@@ -41,7 +41,7 @@ object ChainIndexer {
 
   def apply[F[_]: Sync: Parallel: Timer, D[_]: MonadThrow: LiftConnectionIO](
     settings: IndexerSettings,
-    network: ErgoNetworkClient[F]
+    network: ErgoNetwork[F]
   )(trans: Trans[D, F])(implicit logs: Logs[F, F], makeRef: MakeRef[F, F]): F[ChainIndexer[F]] =
     logs.forService[ChainIndexer[F]].flatMap { implicit log =>
       makeRef.refOf(List.empty[(Id, Int)]).flatMap { updatesRef =>
@@ -53,10 +53,10 @@ object ChainIndexer {
     F[_]: Monad: Parallel: Timer: Bracket[*[_], Throwable]: Logging,
     D[_]: Monad
   ](
-    settings: IndexerSettings,
-    network: ErgoNetworkClient[F],
-    pendingChainUpdates: Ref[F, List[(Id, Int)]],
-    repos: RepoBundle[D]
+     settings: IndexerSettings,
+     network: ErgoNetwork[F],
+     pendingChainUpdates: Ref[F, List[(Id, Int)]],
+     repos: RepoBundle[D]
   )(trans: Trans[D, F])
     extends ChainIndexer[F] {
